IOT-1583: Removing /W3 warnings inside connectivity/ (part 2)
Removing the /W3 warnings in order to add the /WX option to prevent new
ones from being added inside the resource/csdk/connectivity/ directory.

Reviewed-on: https://gerrit.iotivity.org/gerrit/17547
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/csdk/connectivity/inc/caedradapter.h
@@ -122,7 +122,7 @@ int32_t CASendEDRMulticastData(const CAEndpoint_t *endpoint, const void *data,
  *
  * @return ::CA_STATUS_OK or ERROR CODES (::CAResult_t error codes in cacommon.h).
  */
-CAResult_t CAGetEDRInterfaceInformation(CAEndpoint_t **info, uint32_t *size);
+CAResult_t CAGetEDRInterfaceInformation(CAEndpoint_t **info, size_t *size);
 
 /**
  * Read Synchronous API callback.
